requirements.txt
langchain
ipykernel
python-dotenv
# to import all key value pairs from .env apis

langchain_community
# langchain community for all langchain packages

pypdf
# to read a pdf file

bs4
# beautiful soup for web scraping

langchain-text-splitters
# to use all different types of text splitters

python-dotenv
# used to call all the environment variables in the project

langchain-openai
# to deal with openai

chromadb
# to store vectors we use diff dbases one is chromadb in openai

sentence-transformers
# in this library hugging face embedding is available

langchain_huggingface
# langchain and hf are combined to create one library

langchain_chroma
# using chromadb directly from langchain

streamlit
# web app

groq
# to interact with groq apis

langchain_groq
# use langchain with groq

fastapi
# webframework for building apis with python

uvicorn
# webserver to communicate with fastapi

langserve
# to handle deployment , to focus on refining features 

fassi-cpu
# to use fassi
